refactor: split panel handlers modules
Fixes #191
